fix(exec): Fails to Execute Binaries Named After Shell Keywords (#8221)
<!-- What / Why -->
<!-- Describe the request in detail. What it does and why it's being
changed. -->
The best example of this bug is in the original issue (#8190).
Basically, if an exec command has a shell keyword (i.e `npx select`) it
will trigger a bash syntax error.

Things I did:
1. Added a test that fails to execute a binary named after a shell
keyword (`select`)
2. Wrap all execution commands (after handling parsing for
locations/downloads/etc) in quotes to guarantee a programs execution and
not a bash parse

This is my first PR on this project so please let me know if I missed
something! I'm happy to make whatever changes we need to get it fixed!

Thanks! 

## References
Fixes #8190